Update-16.04.19
Corrected the algorithm in X_H2_bar.py
Included efficient Self-Shielding using loops instead of individual variables.
One M, one Z, one G_o, and one n_H_mean for one cell, and so one X_H2_bar for that one cell.

NOTE: This algorithm is the direct copy of the code used in the simulation post processing. Since there is already a grid in the simulation, there is no need to define a grid, i.e., the previous algorithm looped over M, Z, G_o- meaning for one M, there was a grid of values of X_H2_bar and n_H_mean-this was needed to create a grid without the simulation. So, the old algorithm was also correct, but just not what I needed for post processing the simulation. But the old algorithm produced plots that are serving as a base to check whether the plots produced by the simulation are correct or not.