Latest backup of stuff in the notebook on the cluster
Latest backup of stuff in the notebook on the cluster in the folder "outputs". Different versions from 1.1 to 1.5, the info for these versions is also included in a .txt file in the outputs folder.
Also included is the ".sh" file that is the jobscript that has to be submitted to the cluster to start the computation in the background. The name of the .py file that has to run in the background has to be inserted in the script. Also, the time can be changed (max 16:00:00 hours).