Update-19.05.19
Added a separate sub_grid.py file, and the resulting arrays and plots from this file.

This is code just for the sub_grid model, and to save X_H2, X_CO, n_CO, lambda_jeans arrays to use in rad_transfer code later.
Values from simulation are not used here, this is purely theoretical.
pynbody is only imported for assigning units to different quantities, although after importing the saved array the units are not imported for some reason, so importing pynbody is useless. While importing, pynbody can be used to assign the units again.
Also, X_H2_bar and X_CO_bar are calculated and stored in:
     reference_plots => for purely theoretical values
     3.8.py, 4.0.py => for 5kpc and 15kpc versions by using simulation values